backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.0
greenlet==3.1.1
asyncpg==0.30.0
email-validator==2.2.0
bcrypt==4.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
